+++ provider.py
@@ -32,6 +32,7 @@ from .algorithms.alg_performance_report import PlanPerformanceReportAlgorithm
 from .algorithms.alg_facility_location import FacilityLocationAlgorithm
 from .algorithms.alg_capacitated_allocation import CapacitatedAllocationAlgorithm
 from .algorithms.alg_land_allocation import LandUseAllocationAlgorithm
+from .algorithms.alg_pareto_allocation import ParetoAllocationAlgorithm
 from .algorithms.alg_access_equity import AccessEquityAlgorithm
 
 
@@ -86,5 +87,6 @@ class PlanXProvider(QgsProcessingProvider):
         self.addAlgorithm(FacilityLocationAlgorithm())
         self.addAlgorithm(CapacitatedAllocationAlgorithm())
         self.addAlgorithm(LandUseAllocationAlgorithm())
+        self.addAlgorithm(ParetoAllocationAlgorithm())
         # 9 | Equity
         self.addAlgorithm(AccessEquityAlgorithm())